requirements.txt
aiohttp==2.2.3
appnope==0.1.0
async-timeout==1.2.1
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
bleach==2.0.0
boto3==1.4.4
botocore==1.5.78
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cycler==0.10.0
decorator==4.0.11
docutils==0.13.1
entrypoints==0.2.3
future==0.16.0
h5py==2.7.1
html5lib==0.999999999
hyperlink==17.2.1
idna==2.6
incremental==17.5.0
iniherit==0.3.9
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
joblib==0.11
json-rpc==1.10.8
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.3.3
jupyter-core==4.3.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.1
jupyter-nbextensions-configurator==0.3.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.7.4
multidict==3.1.0
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.3
pandas==0.21.1
pandocfilters==1.4.1
patsy==0.4.1
peewee==2.10.2
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.5.2
prompt-toolkit==1.0.14
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
pymc3==3.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-language-server==0.11.1
python-slugify==1.2.4
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.4
rope==0.10.7
s3transfer==0.1.10
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
terminado==0.6
testpath==0.3.1
Theano==1.0.1
tornado==4.5.1
tqdm==4.19.5
traitlets==4.3.2
Twisted==17.5.0
txaio==2.8.0
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
yapf==0.19.0
yarl==0.11.0
yoyo-migrations==5.0.5
zope.interface==4.4.2
